requirements.txt
streamlit>=1.40.0
requests
python-dotenv
